fix: make clippy happy
