+++ crates/rust-client/src/note/note_screener.rs
@@ -16,7 +16,6 @@ use miden_tx::{
 use thiserror::Error;
 
 use crate::{
-    builder::DebugMode,
     store::{Store, StoreError, data_store::ClientDataStore},
     transaction::{TransactionRequestBuilder, TransactionRequestError},
 };
@@ -121,8 +120,10 @@ impl NoteScreener {
         let transaction_request =
             TransactionRequestBuilder::new().build_consume_notes(vec![note.id()])?;
 
-        let tx_script = transaction_request
-            .build_transaction_script(&AccountInterface::from(account), DebugMode::Enabled)?;
+        let tx_script = transaction_request.build_transaction_script(
+            &AccountInterface::from(account),
+            crate::DebugMode::Enabled,
+        )?;
 
         let tx_args = transaction_request.clone().into_transaction_args(tx_script, vec![]);
         let input_notes = InputNotes::new(vec![InputNote::unauthenticated(note.clone())])
